requirements_client.txt
pillow~=10.2.0
requests~=2.31.0
numpy~=1.26.4
